+++ state/cpu_t.go
@@ -4,12 +4,14 @@ type cpu_t struct {
 	Prev           []int // Slice of the previous stat's jiffies.
 	Cur            []int // Slice of the current stat's jiffies.
 	LastCPUIdle    int
-	LastCPUPercent float32
+	LastCPUPercent float64
 	LastSum        int
 }
 
 // Pushes current to previous and sets current to newest
-// file data from /proc/stat.
+// #file data from /proc/stat.
+//
+//	fmt.Printf("$i")
 func (c *cpu_t) Add(nums []int) {
 	c.Prev = c.Cur
 	c.Cur = nums
